Enable gradle on the directory
