+++ src/main/java/duke/commands/ByeCommand.java
@@ -1,4 +1,4 @@
-package commands;
+package duke.commands;
 
 public class ByeCommand extends Command {
     public static final String COMMAND_WORD = "bye";
