+++ src/main/java/duke/commands/CommandResult.java
@@ -1,4 +1,4 @@
-package commands;
+package duke.commands;
 
 /**
  * Represents the result of a command execution.
