+++ src/main/java/duke/commands/IncorrectCommand.java
@@ -1,4 +1,4 @@
-package commands;
+package duke.commands;
 
 
 /**
