+++ src/main/java/duke/common/Messages.java
@@ -1,4 +1,4 @@
-package common;
+package duke.common;
 
 public class Messages {
     public static final String MESSAGE_GOODBYE = "Bye. Hope to see you again soon!";
