+++ src/main/java/duke/exception/MalformedUserInputException.java
@@ -1,4 +1,4 @@
-package exception;
+package duke.exception;
 
 public class MalformedUserInputException extends Exception {
     public MalformedUserInputException(String message) {
