+++ src/main/java/duke/tasklist/Todo.java
@@ -1,4 +1,4 @@
-package tasklist;
+package duke.tasklist;
 
 public class Todo extends Task {
